+++ hardware/EvohomeScript.cpp
@@ -25,7 +25,6 @@
   //#include "../main/RFXtrx.h"
   //#include "../main/Helper.h"
 #include "../main/SQLHelper.h"
-//#include "../main/localtime_r.h"
 #include "../main/mainworker.h"
 
 #include <string>
@@ -146,7 +145,7 @@ void CEvohomeScript::RunScript(const char* pdata, const unsigned char /*length*/
 #endif
 			std::string scriptparams;
 			//Add parameters
-			int pindex = scriptname.find(' ');
+			size_t pindex = scriptname.find(' ');
 			if (pindex != std::string::npos)
 			{
 				scriptparams = scriptname.substr(pindex + 1);
