+++ hardware/KMTronic433.cpp
@@ -3,7 +3,6 @@
 #include "../main/Logger.h"
 #include "../main/Helper.h"
 #include "../main/RFXtrx.h"
-#include "../main/localtime_r.h"
 #include "P1MeterBase.h"
 #include "hardwaretypes.h"
 #include <string>
@@ -147,7 +146,7 @@ void KMTronic433::readCallback(const char *data, size_t len)
 		return; //receiving not enabled
 
 	memcpy(m_buffer, data, len);
-	m_bufferpos = len;
+	m_bufferpos = (int)len;
 }
 
 bool KMTronic433::WriteInt(const unsigned char *data, const size_t len, const bool bWaitForReturn)
