+++ hardware/MySensorsTCP.cpp
@@ -3,12 +3,11 @@
 #include "../main/Logger.h"
 #include "../main/Helper.h"
 #include <iostream>
-#include "../main/localtime_r.h"
 #include "../main/mainworker.h"
 
 #define RETRY_DELAY 30
 
-MySensorsTCP::MySensorsTCP(const int ID, const std::string &IPAddress, const unsigned short usIPPort)
+MySensorsTCP::MySensorsTCP(const int ID, const std::string& IPAddress, const unsigned short usIPPort)
 	: m_retrycntr(RETRY_DELAY)
 	, m_szIPAddress(IPAddress)
 	, m_usIPPort(usIPPort)
@@ -94,9 +93,9 @@ void MySensorsTCP::Do_Work()
 	Log(LOG_STATUS, "TCP/IP Worker stopped...");
 }
 
-void MySensorsTCP::OnData(const unsigned char *pData, size_t length)
+void MySensorsTCP::OnData(const unsigned char* pData, size_t length)
 {
-	ParseData(pData, length);
+	ParseData(pData, (int)length);
 }
 
 void MySensorsTCP::OnError(const boost::system::error_code& error)
@@ -122,7 +121,7 @@ void MySensorsTCP::OnError(const boost::system::error_code& error)
 		Log(LOG_ERROR, "%s", error.message().c_str());
 }
 
-void MySensorsTCP::WriteInt(const std::string &sendStr)
+void MySensorsTCP::WriteInt(const std::string& sendStr)
 {
 	if (!isConnected())
 	{
