+++ hardware/OctoPrintMQTT.cpp
@@ -3,7 +3,6 @@
 #include "../main/Logger.h"
 #include "../main/Helper.h"
 #include <iostream>
-#include "../main/localtime_r.h"
 #include "../main/mainworker.h"
 #include "../main/SQLHelper.h"
 #include "../main/json_helper.h"
@@ -276,7 +275,7 @@ void COctoPrintMQTT::SendMessage(const std::string &Topic, const std::string &Me
 			Log(LOG_STATUS, "Not Connected, failed to send message: %s", Message.c_str());
 			return;
 		}
-		publish(nullptr, Topic.c_str(), Message.size(), Message.c_str());
+		publish(nullptr, Topic.c_str(), (int)Message.size(), Message.c_str());
 	}
 	catch (...)
 	{
